move Politician into own folder
