+++ src/components/Politician/Mobile.vue
