+++ src/router/index.js
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import Router from 'vue-router'
 import HelloWorld from '@/components/HelloWorld'
 import Home from '@/components/Home'
-import Politician from '@/components/Politician'
+import Politician from '@/components/Politician/index'
 import Politicians from '@/components/Politicians'
 import Promise from '@/components/Promise'
 import Promises from '@/components/Promises'
